# api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {

    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:18083",
  "basePath": "/",
...
  "paths": {
...
    "/eventlistener": {
      "post": {
        "tags": [
          "event-listener-controller"
        ],
        "summary": "registEventListner",
        "operationId": "registEventListnerUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "requestEventListenerDto",
            "description": "requestEventListenerDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RequestEventListenerDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseWrapper«long»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/eventlistener/list": {
      "get": {
        "tags": [
          "event-listener-controller"
        ],
        "summary": "getEventListenerList",
        "operationId": "getEventListenerListUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseWrapper«List«ResponseEventListenerDto»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/eventlistener/run/{eventListenerIdx}": {
      "get": {
        "tags": [
          "event-listener-controller"
        ],
        "summary": "runEventListener",
        "operationId": "runEventListenerUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "eventListenerIdx",
            "in": "path",
            "description": "eventListenerIdx",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseWrapper«boolean»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/eventlistener/workflowDetail/{workflowIdx}/{evnetListenerYn}": {
      "get": {
        "tags": [
          "event-listener-controller"
        ],
        "summary": "getWorkflowDetail",
        "operationId": "getWorkflowDetailUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "workflowIdx",
            "in": "path",
            "description": "workflowIdx",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "evnetListenerYn",
            "in": "path",
            "description": "evnetListenerYn",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseWrapper«WorkflowDetailResDto»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/eventlistener/workflowList/{eventListenerYn}": {
      "get": {
        "tags": [
          "event-listener-controller"
        ],
        "summary": "getWorkflowList",
        "operationId": "getWorkflowListUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "eventListenerYn",
            "in": "path",
            "description": "eventListenerYn",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseWrapper«List«WorkflowListResDto»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/eventlistener/{eventListenerIdx}": {
      "get": {
        "tags": [
          "event-listener-controller"
        ],
        "summary": "detailEventListener",
        "operationId": "detailEventListenerUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "eventListenerIdx",
            "in": "path",
            "description": "eventListenerIdx",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseWrapper«ResponseEventListenerDto»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "event-listener-controller"
        ],
        "summary": "deleteEventListner",
        "operationId": "deleteEventListnerUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "eventListenerIdx",
            "in": "path",
            "description": "eventListenerIdx",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseWrapper«boolean»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "patch": {
        "tags": [
          "event-listener-controller"
        ],
        "summary": "updateEventListner",
        "operationId": "updateEventListnerUsingPATCH",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "eventListenerIdx",
            "in": "path",
            "description": "eventListenerIdx",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "requestEventListenerDto",
            "description": "requestEventListenerDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RequestEventListenerDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseWrapper«boolean»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/oss/list/{ossTypeName}": {
      "get": {
        "tags": [
          "oss-controller"
        ],
        "summary": "getOssList",
        "operationId": "getOssListUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "ossTypeName",
            "in": "path",
            "description": "ossTypeName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseWrapper«List«OssDto»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/workflow/list": {
      "get": {
        "tags": [
          "workflow-controller"
        ],
        "summary": "getWorkflowList",
        "operationId": "getWorkflowListUsingGET_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseWrapper«List«WorkflowListResDto»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "WorkflowDto": {
      "type": "object",
      "properties": {
        "ossIdx": {
          "type": "integer",
          "format": "int64"
        },
        "script": {
          "type": "string"
        },
        "workflowIdx": {
          "type": "integer",
          "format": "int64"
        },
        "workflowName": {
          "type": "string"
        },
        "workflowPurpose": {
          "type": "string"
        }
      }
    },
...
    "ResponseWrapper«List«WorkflowListResDto»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/WorkflowListResDto"
          }
        },
        "detail": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "ResponseEventListenerDto": {
      "type": "object",
      "properties": {
        "eventListenerCallUrl": {
          "type": "string"
        },
        "eventListenerDesc": {
          "type": "string"
        },
        "eventListenerIdx": {
          "type": "integer",
          "format": "int64"
        },
        "eventListenerName": {
          "type": "string"
        },
        "eventListenerUrl": {
          "type": "string"
        },
        "workflowIdx": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "ResponseWrapper«List«ResponseEventListenerDto»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ResponseEventListenerDto"
          }
        },
        "detail": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "WorkflowParamDto": {
      "type": "object",
      "properties": {
        "eventListenerYn": {
          "type": "string"
        },
        "paramIdx": {
          "type": "integer",
          "format": "int64"
        },
        "paramKey": {
          "type": "string"
        },
        "paramValue": {
          "type": "string"
        },
        "workflowIdx": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
    "ResponseWrapper«List«OssDto»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OssDto"
          }
        },
        "detail": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "RequestEventListenerDto": {
      "type": "object",
      "properties": {
        "eventListenerDesc": {
          "type": "string"
        },
        "eventListenerIdx": {
          "type": "integer",
          "format": "int64"
        },
        "eventListenerName": {
          "type": "string"
        },
        "eventListenerUrl": {
          "type": "string"
        },
        "workflowIdx": {
          "type": "integer",
          "format": "int64"
        },
        "workflowParams": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/WorkflowParamDto"
          }
        }
      }
    },
    "WorkflowStageMappingDto": {
      "type": "object",
      "properties": {
        "mappingIdx": {
          "type": "integer",
          "format": "int64"
        },
        "stageContent": {
          "type": "string"
        },
        "stageOrder": {
          "type": "integer",
          "format": "int32"
        },
        "workflowIdx": {
          "type": "integer",
          "format": "int64"
        },
        "workflowStageIdx": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
    "ResponseWrapper«boolean»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "boolean"
        },
        "detail": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "ResponseWrapper«long»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "integer",
          "format": "int64"
        },
        "detail": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "OssDto": {
      "type": "object",
      "properties": {
        "ossDesc": {
          "type": "string"
        },
        "ossIdx": {
          "type": "integer",
          "format": "int64"
        },
        "ossName": {
          "type": "string"
        },
        "ossPassword": {
          "type": "string"
        },
        "ossTypeIdx": {
          "type": "integer",
          "format": "int64"
        },
        "ossUrl": {
          "type": "string"
        },
        "ossUsername": {
          "type": "string"
        }
      }
    },
    "WorkflowDetailResDto": {
      "type": "object",
      "properties": {
        "workflowInfo": {
          "$ref": "#/definitions/WorkflowDto"
        },
        "workflowParams": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/WorkflowParamDto"
          }
        },
        "workflowStageMappings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/WorkflowStageMappingDto"
          }
        }
      }
    },
...
    "WorkflowListResDto": {
      "type": "object",
      "properties": {
        "workflowInfo": {
          "$ref": "#/definitions/WorkflowDto"
        },
        "workflowParams": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/WorkflowParamDto"
          }
        },
        "workflowStageMappings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/WorkflowStageMappingDto"
          }
        }
      }
    },
    "ResponseWrapper«ResponseEventListenerDto»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/ResponseEventListenerDto"
        },
        "detail": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "ResponseWrapper«WorkflowDetailResDto»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/WorkflowDetailResDto"
        },
        "detail": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
  }
}